+++ src/components/ListComponent.js
@@ -31,9 +31,22 @@ const getComponent = (query, listTypesSettings) =>
     [QUERY_TYPES.POINTS_OF_INTEREST_AND_TOURS]: getListComponent(
       listTypesSettings[QUERY_TYPES.POINTS_OF_INTEREST_AND_TOURS]
     ),
+    [QUERY_TYPES.GENERIC_ITEMS]: TextList,
     [QUERY_TYPES.CATEGORIES]: CategoryList
   }[query]);
 
+const isHorizontal = (query, listTypesSettings) => {
+  switch (query) {
+    case QUERY_TYPES.TOURS:
+    case QUERY_TYPES.POINTS_OF_INTEREST:
+      return listTypesSettings[QUERY_TYPES.POINTS_OF_INTEREST_AND_TOURS] === LIST_TYPES.CARD_LIST;
+    default:
+      return listTypesSettings[query] === LIST_TYPES.CARD_LIST;
+  }
+};
+
+// the ListComponent will default to being horizontal for CardLists,
+// which can be overwritten by passing in the horizontal prop
 export const ListComponent = ({
   navigation,
   data,
@@ -52,7 +65,7 @@ export const ListComponent = ({
     <Component
       data={data}
       fetchMoreData={fetchMoreData}
-      horizontal={horizontal}
+      horizontal={horizontal ?? isHorizontal(query, listTypesSettings)}
       ListHeaderComponent={ListHeaderComponent}
       navigation={navigation}
       noSubtitle={noSubtitle}